+++ lib/eve_online/esi/universe_star.rb
@@ -7,6 +7,8 @@ module EveOnline
     class UniverseStar < Base
       extend Forwardable
 
+      include ResponseWithEtag
+
       API_ENDPOINT = 'https://esi.evetech.net/v1/universe/stars/%<star_id>s/?datasource=%<datasource>s'
 
       attr_reader :id
@@ -18,10 +20,11 @@ module EveOnline
       end
 
       def_delegators :model, :as_json, :age, :luminosity, :name, :radius,
-                     :solar_system_id, :spectral_class, :temperature, :type_id
+                     :solar_system_id, :spectral_class, :temperature,
+                     :type_id, :etag
 
       def model
-        @model ||= Models::Star.new(response)
+        @model ||= Models::Star.new(response_with_etag)
       end
 
       def scope; end